Move implementation of hasPermittedSubclasses from Classes to AbstractClassAssert
